requirements.txt
pandas==0.24.2
transformers>=4.2.0
matplotlib>=3.3.2
numpy>=1.18.5
pytorch-lightning>=0.9.0
PyYAML>=5.3.1
ray>=0.8.7
scikit-learn>=0.23.2
scipy>=1.5.2
tensorboard>=2.2.0
torch>=1.6.0
torch-cluster>=1.5.7
torch-geometric>=1.6.1
torch-scatter>=2.0.5
torch-sparse>=0.6.7
torch-spline-conv>=1.2.0
torchvision>=0.7.0
tqdm>=4.49.0
